requirements.txt
dbt-core==1.1.1
dbt-snowflake==1.1.0
streamlit==1.12.0
